class IpForwardingUtils implemented in distro_lib
The old implementation of IpForwardingUtils uses ip route command, but
this is not suported in FreeBSD. This commit moves the implementation to
distro_lib and transforms IpForwardingUtils into an interface class,
allowing each distro to chose which implementation to use.

class IpForwardingUtilsIprouteTest:
	contains the old implemetation using ip route command
class IpForwardingUtilsIfconfig:
	contains an implementation using ifconfig command
