+++ google_compute_engine/distro_lib/freebsd_11/utils.py
@@ -16,6 +16,7 @@
 """Utilities that are distro specific for use on FreeBSD 11."""
 
 from google_compute_engine.distro_lib import helpers
+from google_compute_engine.distro_lib import ip_forwarding_utils
 from google_compute_engine.distro_lib import utils
 
 
@@ -40,3 +41,12 @@ class Utils(utils.Utils):
       logger: logger object, used to write to SysLog and serial port.
     """
     helpers.CallNtpdate(logger)
+
+  def IpForwardingUtils(self, logger, proto_id=None):
+    """Get system IP address configuration utilities.
+
+    Args:
+      logger: logger object, used to write to SysLog and serial port.
+      proto_id: string, the routing protocol identifier for Google IP changes.
+    """
+    return ip_forwarding_utils.IpForwardingUtilsIfconfig(logger)
