+++ google_compute_engine/distro_lib/sles_11/utils.py
@@ -20,6 +20,7 @@ import subprocess
 
 from google_compute_engine import constants
 from google_compute_engine.distro_lib import helpers
+from google_compute_engine.distro_lib import ip_forwarding_utils
 from google_compute_engine.distro_lib import utils
 
 
@@ -67,3 +68,12 @@ class Utils(utils.Utils):
       logger: logger object, used to write to SysLog and serial port.
     """
     helpers.CallHwclock(logger)
+
+  def IpForwardingUtils(self, logger, proto_id=None):
+    """Get system IP address configuration utilities.
+
+    Args:
+      logger: logger object, used to write to SysLog and serial port.
+      proto_id: string, the routing protocol identifier for Google IP changes.
+    """
+    return ip_forwarding_utils.IpForwardingUtilsIproute(logger, proto_id)
