+++ google_compute_engine/distro_lib/utils.py
@@ -45,3 +45,12 @@ class Utils(object):
       logger: logger object, used to write to SysLog and serial port.
     """
     pass
+
+  def IpForwardingUtils(self, logger, proto_id=None):
+    """Get system IP address configuration utilities.
+
+    Args:
+      logger: logger object, used to write to SysLog and serial port.
+      proto_id: string, the routing protocol identifier for Google IP changes.
+    """
+    pass
