+++ google_compute_engine/networking/tests/network_daemon_test.py
@@ -149,8 +149,8 @@ class NetworkDaemonTest(unittest.TestCase):
     expected_calls = [
         mock.call.setup._ExtractInterfaceMetadata(result),
         mock.call.network_setup.EnableNetworkInterfaces(['a', 'b']),
-        mock.call.forwarding.HandleForwardedIps('a', None),
-        mock.call.forwarding.HandleForwardedIps('b', None),
+        mock.call.forwarding.HandleForwardedIps('a', None, None),
+        mock.call.forwarding.HandleForwardedIps('b', None, None),
     ]
     self.assertEqual(mocks.mock_calls, expected_calls)
